backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0
greenlet==3.0.3

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # passlib 1.7.4 不兼容 bcrypt>=4.1

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# 工具库
cryptography==41.0.7
pycryptodome==3.20.0
dnspython==2.5.0
whois==0.9.27
requests==2.31.0
httpx==0.28.1

# 缓存
cachetools==5.3.2

# 其他
python-dateutil==2.8.2
aiofiles==23.2.1

# 中文分词
jieba==0.42.1

